CMenuWidget: add member to suppress paint of details line
In some situations is it more senseful to suppress painting of details line.
Required if menu window is painted over another window.
The hide() functionality of detailsline paints only an empty background on screen so we have
blank parts on screen. That looks bad. For example, this is to observe
on context menu in channellist.

We should see this as a workaround. It is not certain whether the
details line is really needed in the future.

Conflicts:
	src/gui/widget/menue.cpp